Code resturcturing to allow for init flag and terminal app to draft configuration file, module import
